+++ src/main/java/com/bobby/artistweb/model/UniqueValues.java
@@ -14,6 +14,6 @@ public class UniqueValues {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id = 0;
+    private long id = 0;
     private String value;
 }
