+++ src/main/java/com/bobby/artistweb/repo/ForegroundImageRepo.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ForegroundImageRepo extends JpaRepository<ForegroundImage, Integer> {
+public interface ForegroundImageRepo extends JpaRepository<ForegroundImage, Long> {
 }
